tegra: work in progress jp 5.1 EA changes
Put this on top of the parent commit for [1] before
we made the switch to tegra-bootloader-update on Jetpack 4.

Attempt to prove we can run nv_update_engine to successfully update
bootloader partitions in Jetpack 5.1 EA.

* Remove logic to support A/B partitions for the rootfs, as this
is not ready to test yet, just bind mount the rootfs to the tmpfs
mount directory.
* Add mount partitions for efi support

1: https://github.com/mendersoftware/meta-mender-community/commit/b5243eb458e6ba824e595febaf07f523cc4241f8