tegra: disable tegra-boot-tools with uboot builds
As noted in the thread at [1], when uboot is used as the bootloader
on t186 platform, the redundant uboot variables are corrupted by
the tegra-bootinfo utility.

Since tegra-bootinfo should not be needed for uboot builds, the
safest fix is likely to just remove these from the build.

If tegra-bootinfo is needed/desired on these builds, an alternate
solution would be to either move or reduce the size of the uboot
environment.

1: https://matrix.to/#/!mKaltPMAHUgQwPUKcH:gitter.im/$jXeFhmQ_6jnmj4B9FwfB2nMObSkU---TnmR4Rgwp33s?via=gitter.im&via=matrix.org&via=3dvisionlabs.com