tegra: fix u-boot variable storage on t186
In the latest r32.x releases of Mender (including dunfell and
kirkstone-l4t-r32.x) there's a conflict between the storage
location used to write bootinfo with the tegra-bootinfo tool at
[1] and the uboot environment which we've attempted to keep
at the same location as previous mender releases.  Note that
this only applies to targets using the cboot instead of the
u-boot bootloader, and the cboot bootloader was the default
for mender integration in recent releases.  So the only likely
issues will be for people who are upgrading from a previous
release which used uboot as the default.

This results in a case where the redundant u-boot environment
is always corrupted at boot by the tegra-bootinfo tool, which
attempts to write successful boot information to extended
storage location starting at 0x3F7800.

This is difficult to catch in the typical mender configuration
because the mender client typically writes variables from
userspace shortly after boot, and recovers the redundant variable
partition as soon as the variables are written.  The easiest way
to reproduce is to do something like:
```
rm -f etc/systemd/system/multi-user.target.wants/mender-client.service
```
to prevent mender client from running at startup.

After doing this, you will notice
```
fw_printenv -n mender_boot_part
Cannot read environment, using default
```
on attempts to read any variable.

There are two options supported here to resolve, and the best
option will likely be related to the state of any current mender
deployments.

The first option, and default in this commit, is to set
```
TEGRA_MENDER_UBOOT_T186_CONFIG_ENV_SIZE_REDUCE = "1"
```
based on a default setting which checks for u-boot bootloader
configuration and ensures the second workaround is not in use.

This will reduce the overall size of the uboot environment space
to 0x1C0000, which means the redundant storage location ends at
0x3F7000 and does not overlap tegra-bootinfo extended sector
storage with build defaults.

This would be an appropriate option to resolve if you aren't using
u-boot storage for any environment variables outside of uboot
defaults, or your total amount of additional variable storage
is negigable, which is likely the case in most scenarios.  It also
has the least impact to existing installations since the redundant
partition is already corrupt due to the issue described here.
However, since fw_env and uboot binaries are not updated on a
mender update there will still be issues with this one on
existing releases

The second option supported by this commit is
```
TEGRA_MENDER_BOOTINFO_STORAGE_REDUCE = "1"
```
which modifies tegra-bootinfo to use only a single extended
partition for storage, thereby avoiding overwriting the default
redundant uboot environment storage location.  Since the only
storage used is less than one sector, assuming you aren't storing
any additional variables beyond defaults (storage of the machine-id
for systemd) there should be no issue with storage size.  However,
in an existing installation, the tegra-bootinfo extended storage
will be corrupt after installing this upgrade, so this upgrade
path would potentially result in losing the machine-id for any
existing installations.  However, for an initial deployment
on t186 this would be an option which would still keep the full
default size of the uboot environment available

Only one of these two options should be selected.  Selecting

`TEGRA_MENDER_BOOTINFO_STORAGE_REDUCE = "1"` will automatically
deselect the default `TEGRA_MENDER_UBOOT_T186_CONFIG_ENV_SIZE_REDUCE = "1"`
enabled with this commit.

See relevant element thread at [2] for additional background
and discussion.

1: https://github.com/OE4T/tegra-boot-tools/blob/master/doc/tegra-bootinfo.md
2: https://matrix.to/#/!mKaltPMAHUgQwPUKcH:gitter.im/$xvPfnhEYNz4OWlgpOsM4v_3ySpLkJ6xRFocjj0G_7eI?via=gitter.im&via=matrix.org&via=3dvisionlabs.com